Generalize LogisticRegression over Floats
This commit generalizes the logistic regression code for floats
so that clients can work with both f32 and f64.

Unfortunately, to make the this work for Argmin, we have to introduce
newtypes for the Floats and Arrays passed to Argmin and then re-implement
a lot of the traits to make this all work.