Add two-class logistic regression (#24)
* Add two-class logistic regression

A simple two-class logistic regression classifier, using argmin's
LBFGS for minimization. The implementation was largely inspired
by Python's sklearn logistic regression.

* Allow users to supply an initial parameter guess

* Generalize LogisticRegression over Floats

This commit generalizes the logistic regression code for floats
so that clients can work with both f32 and f64.

Unfortunately, to make the this work for Argmin, we have to introduce
newtypes for the Floats and Arrays passed to Argmin and then re-implement
a lot of the traits to make this all work.

* Add comments to explain purpose of argmin_helper

* Add test case for working with f32's

* Generalize logistic regression over Array types

This change allows the main methods of logistic regression to be called
with any kind of array from ndarray - not just with owned array
representations. This makes the interface more flexible to callers.

* Remove FloatWrapper

After some help from @bytesnake, FloatWrapper isn't actually required. This
saves a lot of boilerplate.

* Fix tests by requiring less precise results

* Make labels more flexible and return probabilities

Allow referencing iterators over all types which implement
`PartialEq + Clone` to make the fit method as flexible as possible.

Provide two separate methods `predict_probabilities` and
`predict_classes`.

Allow setting a different probability threshold for when to predict
a given class.