Feature/optics (#58)
* oh hey it's optics again

* rename optics result

* Fix merge snafu

* Apply some PR feedback

* Move to use HNSW for nearest neighbour calculation

* Get rid of processed bool

* Avoid going over all the points again and again

Use a loop similar to PR feedback

* Change to std::f64::MAX for MSRV in CI

* Update Hnsw and add suggested comment

* Update hnsw again

* Update comment and move seeds creation

* Move to error and validate pattern

* Move algorithm to return result type

* Add some tests

Tests also highlighted some mistakes made during refactoring that broke
some parts.

* Test find_neighbors and get_seeds

* Up MSRV

* Apply some review comments

* Remove FloatOrd and replace with noisy_float

* Remove lifetimes from OpticsAnalysis

Remove the observation field from samples keeping just the index into
the data, removing the lifetime faff. Also renamed minimum_points to
get_minimum_points. Next step is simplifying the trait impl to just be
on ArrayView

* Remove observation data from neighbours

Needed to fix lifetime issues with the changes for OpticsAnalysis

* Fix tests

* Update CI to work with min-const generics

* Update to actual min const generics stablisation version

* Fix coverage

* Move Analysis over to another file

The benefit of this is that other algorithms can be implemented on the
analysis results, so by moving it out we make it tidier to implement
things using the analysis in future. Also make the fields non-public and
implement some methods to access data

* update MSRV for test

* Apply PR feedback

* Update docs from feedback

* Add optics example

* Set tolerance

* Add script

* Apply PR suggestion